教学评价

+++ src/main/java/com/mycode/jiaoxuepingjia/pjset/service/PjSetServiceImpl.java
@@ -2,7 +2,6 @@ package com.mycode.jiaoxuepingjia.pjset.service;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
-import com.mycode.jiaoxuepingjia.pjset.domain.PjSet;
 import com.mycode.jiaoxuepingjia.pjset.domain.PjSetTarget;
 import com.mycode.jiaoxuepingjia.pjset.domain.PjSetTemplate;
 import com.mycode.jiaoxuepingjia.pjset.mapper.PjSetMapper;
@@ -24,12 +23,17 @@ public class PjSetServiceImpl implements PjSetService {
     private PjSetMapper pjSetMapper;
 
     @Override
-    public Map<String, Object> getPjSetList(PjSet pjSet) {
+    public Map<String, Object> getExecTemplate(String templateType) {
         Map<String, Object> resultMap = new HashMap<>();
-        Page<Object> pageInfo = PageHelper.startPage(pjSet.getPageIndex(), pjSet.getPageSize());
-        List<PjSet> pageList = pjSetMapper.getPjSetList(pjSet);
-        resultMap.put("totalNum", pageInfo.getTotal());
-        resultMap.put("pageList", pageList);
+        String templateCode = pjSetMapper.getExecTemplate(templateType); //是否评教时间（即查看当前是否有执行中的模板信息）
+        boolean isPjDate = false;
+        List<PjSetTarget> pjSetTargetList = null;
+        if(StringUtils.isNotEmpty(templateCode)){
+            isPjDate = true;
+            pjSetTargetList = pjSetMapper.getPjSetTargetListByTemplateCode(templateCode); //根据模板编号获取模板信息
+        }
+        resultMap.put("isPjDate",isPjDate);
+        resultMap.put("targetList",pjSetTargetList);
         return resultMap;
     }
 
@@ -44,26 +48,6 @@ public class PjSetServiceImpl implements PjSetService {
     }
 
     @Override
-    public List<PjSetTarget> getPjSetTargetList(PjSetTarget pjSetTarget) {
-        return pjSetMapper.getPjSetTargetList(pjSetTarget);
-    }
-
-    @Override
-    public Map<String, Object> getCurrentTemplate(String templateType) {
-        Map<String, Object> resultMap = new HashMap<>();
-        String templateCode = pjSetMapper.isPj(templateType); //查看当前是否有可用的模板信息
-        boolean isPj = false;
-        List<PjSetTarget> pjSetTargetList = null;
-        if(StringUtils.isNotEmpty(templateCode)){
-            isPj = true;
-            pjSetTargetList = pjSetMapper.getPjSetTargetListByTemplateCode(templateCode); //根据模板编号获取模板信息
-        }
-        resultMap.put("isPj",isPj);
-        resultMap.put("targetList",pjSetTargetList);
-        return resultMap;
-    }
-
-    @Override
     public Boolean insertOrUpdateTemplate(PjSetTemplate template, String[] targetCodes) {
         Boolean bool = false;
         if(template != null && StringUtils.isEmpty(template.getTemplateCode())){
@@ -72,7 +56,7 @@ public class PjSetServiceImpl implements PjSetService {
         }else{
             bool = pjSetMapper.updateTemplate(template);
         }
-        if(bool){ //
+        if(bool && targetCodes !=null){ //
             List<PjSetTarget> targetList = pjSetMapper.getPjSetTargetListByTemplateCode(template.getTemplateCode());
             if(targetList !=null && targetList.size() >0){
                 bool = pjSetMapper.deleteTemplateTargetByTemplateCode(template.getTemplateCode());
@@ -86,6 +70,11 @@ public class PjSetServiceImpl implements PjSetService {
     public Boolean deleteTemplate(String templateCode) {
         Integer execNum = pjSetMapper.deleteTemplate(templateCode); //mybatis一次对多条数据进行操作成功后返回值为 -1
         return execNum < 0;
+    }
+
+    @Override
+    public List<PjSetTarget> getPjSetTargetList(PjSetTarget pjSetTarget) {
+        return pjSetMapper.getPjSetTargetList(pjSetTarget);
     }
 
     @Override
